requirements.txt
alembic==1.17.2
certifi==2025.11.12
charset-normalizer==3.4.4
class-resolver==0.7.1
click==8.3.1
click-default-group==1.2.4
colorama==0.4.6
colorlog==6.10.1
dataclasses-json==0.6.7
docdata==0.0.5
filelock==3.20.0
fsspec==2025.10.0
greenlet==3.2.4
gudhi==3.11.0
idna==3.11
igraph==1.0.0
jinja2==3.1.6
joblib==1.5.2
mako==1.3.10
markupsafe==3.0.3
marshmallow==3.26.1
more-click==0.1.3
more-itertools==10.8.0
mpmath==1.3.0
mypy-extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pip==21.2.3
pykeen==1.11.1
pystow==0.7.11
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
setuptools==57.4.0
six==1.17.0
sqlalchemy==2.0.44
sympy==1.14.0
tabulate==0.9.0
texttable==1.7.0
threadpoolctl==3.6.0
tomli==2.3.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspect==0.9.0
tzdata==2025.2
urllib3==2.5.0
